No inheritance relationship between Pre and Post ContingencyResult (code sharing)
